# openapi.yaml
openapi: 3.0.0
info:
  title: PB138 Study Planner API
  description: API specification for the PB138 Study Planning Application
  version: 1.0.0
  contact:
    name: PB138 Project
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pb138.example.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Subject:
      type: object
      required:
        - id
        - userId
        - name
        - teacher
        - code
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the subject
          example: 42
        userId:
          type: integer
          format: int64
          description: ID of the user who owns this subject
          example: 1
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan
          example: 5
        name:
          type: string
          description: Subject name
          example: "Linear Algebra"
        teacher:
          type: string
          description: Teacher's name
          example: "Prof. Smith"
        code:
          type: string
          description: Subject code (uppercase)
          example: "LA101"
          pattern: "^[A-Z0-9]+$"
        isShared:
          type: boolean
          description: Whether the subject is shared with other users
          default: false
          example: false
        archived:
          type: boolean
          description: Whether the subject is archived (soft deleted)
          example: false
        files:
          type: integer
          description: Count of files associated with this subject
          example: 3
        tasks:
          type: integer
          description: Count of tasks associated with this subject
          example: 5
        events:
          type: integer
          description: Count of events associated with this subject
          example: 2
        notes:
          type: integer
          description: Count of lesson notes associated with this subject
          example: 4
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-05-06T14:20:00.000Z"
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Soft delete timestamp (null if not deleted)
          example: null
    CreateSubjectRequest:
      type: object
      required:
        - name
        - teacher
        - code
      properties:
        name:
          type: string
          minLength: 1
          description: Subject name
          example: "Linear Algebra"
        teacher:
          type: string
          minLength: 1
          description: Teacher's name
          example: "Prof. Smith"
        code:
          type: string
          minLength: 1
          description: Subject code (will be uppercase)
          example: "LA101"
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan
          example: 5
        isShared:
          type: boolean
          description: Whether to share with other users
          default: false
          example: false
    UpdateSubjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          example: "Linear Algebra II"
        teacher:
          type: string
          minLength: 1
          example: "Prof. Johnson"
        code:
          type: string
          minLength: 1
          example: "LA102"
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          example: 5
        isShared:
          type: boolean
          example: true
        archived:
          type: boolean
          description: Set to true to archive, false to unarchive
          example: false
    PaginatedSubjects:
      type: object
      required:
        - data
        - hasMore
        - nextCursor
        - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Subject'
        hasMore:
          type: boolean
          example: false
        nextCursor:
          type: string
          nullable: true
          example: "100"
        limit:
          type: integer
          example: 20
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Neplatne ID ukolu."
security:
  - bearerAuth: []
paths:
...
  /api/subjects:
    get:
      summary: List subjects with optional filtering and pagination
      description: |
        Retrieve subjects. Public users can only see shared subjects.
        Registered users can see their own subjects, shared subjects, and subjects in shared study plans.
      tags:
        - Subjects
      operationId: listSubjects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination (ID of the last item from previous page)
        - name: paginated
          in: query
          schema:
            type: boolean
          description: Enable pagination mode (true/false)
        - name: studyPlanId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by study plan ID
        - name: includeDeleted
          in: query
          schema:
            type: boolean
          description: Include archived subjects (default false)
      responses:
        '200':
          description: List of subjects (with or without pagination)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Subject'
                    description: Regular list when pagination is not enabled
                  - $ref: '#/components/schemas/PaginatedSubjects'
                    description: Paginated response when pagination is enabled
              examples:
                unpaginated:
                  value:
                    - id: 42
                      userId: 1
                      studyPlanId: 5
                      name: "Linear Algebra"
                      teacher: "Prof. Smith"
                      code: "LA101"
                      isShared: false
                      archived: false
                      deletedAt: null
                      files: 3
                      tasks: 5
                      events: 2
                      notes: 4
                      createdAt: "2026-01-15T10:30:00.000Z"
                      updatedAt: "2026-05-06T14:20:00.000Z"
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new subject
      description: |
        Create a subject. If studyPlanId is specified, the authenticated user must be
        the plan owner or a contributor to create subjects in it.
      tags:
        - Subjects
      operationId: createSubject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubjectRequest'
            examples:
              simple:
                summary: Simple subject
                value:
                  name: "Linear Algebra"
                  teacher: "Prof. Smith"
                  code: "LA101"
              withPlan:
                summary: Subject in a study plan
                value:
                  name: "Linear Algebra"
                  teacher: "Prof. Smith"
                  code: "LA101"
                  studyPlanId: 5
                  isShared: true
      responses:
        '201':
          description: Subject created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - no permission to create in this study plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Study plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/subjects/{id}:
    put:
      summary: Update a subject
      tags:
        - Subjects
      operationId: updateSubject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Subject ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubjectRequest'
            example:
              name: "Linear Algebra II"
              code: "LA102"
              isShared: true
      responses:
        '200':
          description: Subject updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Bad request - invalid ID or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a subject (soft delete with cascade)
      description: |
        Soft-deletes a subject and cascades the deletion to:
        - Associated tasks
        - Associated file records
        - Associated lessons
        - Associated events
      tags:
        - Subjects
      operationId: deleteSubject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Subject ID
      responses:
        '200':
          description: Subject deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request - invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'